+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- [Fabric] Fixed death event not working properly
+
 ## [v1.21.11-0.1.3] - 2026-02-08
 
 ### Fixed